+++ global.js
@@ -1,6 +1,6 @@
 import { updateGrid } from "./module/view/GridView.js";
-import { updatePressNewsSection } from "./module/view/ListView/Actions/UpdateSection.js";
-import { changePageInfo, changeCategory } from "./module/view/ListView/Actions/ChangePress.js";
+// import { updatePressNewsSection } from "./module/view/ListView/Actions/UpdateSection.js";
+import { changePageInfo, changeCategory, changePressNewsSection } from "./module/view/ListView/Actions/ChangePress.js";
 
 // 리스트뷰 현재 페이지 & 현재 카테고리
 export const LIST_PAGE = {
@@ -11,7 +11,7 @@ export const LIST_PAGE = {
     LIST_PAGE.CURRENT_PAGE = page;
     // 부분 렌더링
     changePageInfo();
-    updatePressNewsSection();
+    changePressNewsSection();
   },
   setCategory: (category) => {
     LIST_PAGE.CURRENT_CATEGORY = category;
